% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_approx_rxcui.R
\name{find_approx_rxcui}
\alias{find_approx_rxcui}
\title{Find active approximate terms and return their RxCUI}
\usage{
find_approx_rxcui(string, n = 1, minscore = 10, local_host = FALSE)
}
\arguments{
\item{string}{String to search for.}

\item{n}{Number of RxCUI to return}

\item{minscore}{The minimum similarity index for a result to be returned.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
The RxCUI; \code{NULL} if not successful.
}
\description{
See https://lhncbc.nlm.nih.gov/RxNav/news/RxNormApproxMatch.html.
}
\examples{
find_approx_rxcui("FENTANYL 10MCG/ML SOLN INJ PCA 30ML")
}
