% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ndc.R
\name{from_ndc}
\alias{from_ndc}
\title{Find Active RxNorm From NDC}
\usage{
from_ndc(ndc, local_host = FALSE)
}
\arguments{
\item{ndc}{An NDC code as a character string.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
If found, the active RxNorm RxCUI. Otherwise, \code{NA}.
}
\description{
Find Active RxNorm From NDC
}
\examples{
from_ndc("00002143301")
}
