% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dose_form.R
\name{get_dose_form}
\alias{get_dose_form}
\title{Get Dose Form From RxCUI}
\usage{
get_dose_form(rx_cui, local_host = FALSE)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
The dose form; \code{NULL} if not successful.
}
\description{
Get Dose Form From RxCUI
}
\examples{
get_dose_form(1011485)
}
