% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndc_status.R
\name{get_ndc_status}
\alias{get_ndc_status}
\title{Find NDC status code}
\usage{
get_ndc_status(ndc, local_host = FALSE)
}
\arguments{
\item{ndc}{An NDC code as a character string.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
NDC status code.
}
\description{
Find NDC status code
}
\examples{
get_ndc_status("00002143301")
}
