% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rxcui_status.R
\name{get_quantified_rxcui}
\alias{get_quantified_rxcui}
\title{Get New RxCUI from Non-dispensable RxCUI}
\usage{
get_quantified_rxcui(rx_cui, local_host = FALSE)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
If found, the first related RxNorm RxCUI with a quantified factor. Otherwise, \code{NA}.
}
\description{
Get New RxCUI from Non-dispensable RxCUI
}
\examples{
get_quantified_rxcui(1734477)
}
