% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rxcui_status.R
\name{get_remapped_rxcui}
\alias{get_remapped_rxcui}
\title{Get New RxCUI from Remapped RxCUI}
\usage{
get_remapped_rxcui(rx_cui, local_host = FALSE)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
If found, the remapped RxNorm RxCUI. Otherwise, \code{NA}.
}
\description{
Get New RxCUI from Remapped RxCUI
}
\examples{
get_remapped_rxcui(197523)
}
