% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_rx}
\alias{get_rx}
\title{Get Drug Names From RxCUI}
\usage{
get_rx(rx_cui)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}
}
\value{
A list of class \code{rxnorm} containing the following components:

\item{name}{The drug name; \code{NULL} if not successful.}
\item{id}{The RxNorm RxCUI; \code{NULL} if not successful.}
\item{url}{The url used for accessing the REST API.}
}
\description{
Get Drug Names From RxCUI
}
\examples{
get_rx(1011485)
}
