% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rxcui_status.R
\name{get_rxcui_status}
\alias{get_rxcui_status}
\title{Get Status Code for a RxCUI}
\usage{
get_rxcui_status(rx_cui, local_host = FALSE)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
The RxCUI status code.
}
\description{
Get Status Code for a RxCUI
}
\examples{
get_rxcui_status(861819)
}
