% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strength.R
\name{get_rxcui_strength}
\alias{get_rxcui_strength}
\title{Get Strength of Active Ingredients for a RxCUI}
\usage{
get_rxcui_strength(rx_cui, local_host = FALSE)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
A \code{data.frame} of strength information.
}
\description{
Get Strength of Active Ingredients for a RxCUI
}
\examples{
get_rxcui_strength(861819)
}
