% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_who.R
\name{get_who}
\alias{get_who}
\title{Parse WHO ATC/DDD Drug Class}
\usage{
get_who(
  atc,
  query = c("first", "second", "third", "fourth"),
  subsetted = FALSE
)
}
\arguments{
\item{atc}{An WHO ATC/DDD ATC code.}

\item{query}{Level to parse ATC code at. Options are "first" (default),
"second", "third", "fourth".}

\item{subsetted}{Has ATC/DDD query level subsetting already occurred? The default is FALSE.}
}
\value{
The parsed ATC code.
}
\description{
Parse WHO ATC/DDD Drug Class
}
\examples{
get_who("R06AE")
}
