% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{search_rxcui}
\alias{search_rxcui}
\title{Obtain Drug Names From the NIM RxNorm API}
\usage{
search_rxcui(rxcui)
}
\arguments{
\item{rxcui}{Either a string or numeric RxNorm RxCUI to search for.}
}
\value{
A list of class \code{rxnorm} containing the following components:

\item{name}{The drug name; \code{NULL} if not successful.}
\item{id}{The RxNorm RxCUI; \code{NULL} if not successful.}
\item{url}{The url used for accessing the REST API.}
}
\description{
Obtain Drug Names From the NIM RxNorm API
}
\examples{
search_rxcui(1011485)
}
