% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_who.R
\name{subset_atc}
\alias{subset_atc}
\title{Subset WHO ATC Codes}
\usage{
subset_atc(atc, query = c("first", "second", "third", "fourth"))
}
\arguments{
\item{atc}{WHO ATC code.}

\item{query}{The level to subset the code at. Options are "first" (default), "second", "third", "fourth".}
}
\value{
The subsetted code.
}
\description{
Subset ATC codes into the minimal code based on the level of query.
}
\examples{
subset_atc("R06AE", "first")
subset_atc("R06AE", "second")
}
