% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{from_ndc}
\alias{from_ndc}
\title{Find Active RxNorm From NDC}
\usage{
from_ndc(ndc)
}
\arguments{
\item{ndc}{An NDC code as a character string.}
}
\value{
If found, the active RxNorm RxCUI. Otherwise, \code{NA}.
}
\description{
Find Active RxNorm From NDC
}
\examples{
from_ndc("00002143301")
}
