% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_atc}
\alias{get_atc}
\title{Get WHO ATC/DDD Drug Class From RxCUI}
\usage{
get_atc(rx_cui, query_atc = c("none", "main", "thera", "pharma", "chem"))
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{query_atc}{TODO}
}
\value{
A list of class \code{rxnorm} containing the following components:

\item{name}{The brand name(s); \code{NULL} if not successful or not applicable.}
\item{id}{The RxNorm RxCUI(s) for the brand name; \code{NULL} if not successful or not applicable.}
\item{url}{The url used for accessing the REST API.}
}
\description{
Get WHO ATC/DDD Drug Class From RxCUI
}
\examples{
get_atc(1011485)
}
