% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_atc.R
\name{get_atc}
\alias{get_atc}
\title{Get WHO ATC/DDD Drug Class From RxCUI}
\usage{
get_atc(
  rx_cui,
  query_atc = c("none", "first", "second", "third", "fourth"),
  local_host = FALSE
)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{query_atc}{Level to parse ATC code at. Options are "none" (default), "first", "second", "third", "fourth".}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
If \code{query_atc} is "none", the raw ATC code(s), otherwise the
  parsed ATC code(s); \code{NULL} if not successful.
}
\description{
Get WHO ATC/DDD Drug Class From RxCUI
}
\examples{
get_atc(861819)
get_atc(6809)
get_atc(861819, "first")
get_atc(861819, "second")
get_atc(861819, "third")
get_atc(861819, "fourth")
}
