% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_bn}
\alias{get_bn}
\title{Get Drug Brand Names From the NLM RxNorm API}
\usage{
get_bn(rx_cui)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}
}
\value{
A list of class \code{rxnorm} containing the following components:

\item{name}{The brand name(s); \code{NULL} if not successful or not applicable.}
\item{id}{The RxNorm RxCUI(s) for the brand name; \code{NULL} if not successful or not applicable.}
\item{url}{The url used for accessing the REST API.}
}
\description{
Get Drug Brand Names From the NLM RxNorm API
}
\examples{
get_bn(1011485)
get_bn(7052)
}
