% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rx.R
\name{get_rx}
\alias{get_rx}
\title{Get Drug Names From RxCUI}
\usage{
get_rx(rx_cui, local_host = FALSE)
}
\arguments{
\item{rx_cui}{Either a string or numeric RxNorm RxCUI to search for.}

\item{local_host}{Run query locally using RxNav-in-a-box with Docker?}
}
\value{
The drug name; \code{NULL} if not successful.
}
\description{
Get Drug Names From RxCUI
}
\examples{
get_rx(1011485)
}
